Moved bluetooth related class files from internal to internal.bluetooth
